Added NoPage component to display redirect banners
Defined a NoPage.jsx React component to show banners to the user when they try to
access a page which either does not exist or requires a different authentication
status. I tried to change index/routes.jsx to check for a valid token and decide
whether to display NoPage based on that, but there's a delay causing it to
perceive an invalid token right after login. I'll have to rethink this and likely
use the ProtectedRoute component for checking whether the token is valid.