fix: Clean routing index implementation
Updated the routing index for the following functionality:
- Show a new NotFoundPage component for non-existent pages
- Show a new InvaildPage component when an unauthenticated user
  attempts to access a hidden page (e.g. dashboard).

Note: It is not possible to dynamically update the routing index
(list of known paths) to conditionally include hidden pages based
on whether the user holds a token, because this does not update
quickly enough on login.

Instead, we unconditionally include all pages for unauthenticated
users and leave it up to the ProtectedRoute component to ensure
that they are redirected if they try to access a hidden page.